wip(feat): Basic table syntax support
